fix: resolve conversations
