+++ src/app/store.ts
@@ -44,5 +44,7 @@ const store = configureStore({
         }),
 })
 
+export type RootState = ReturnType<typeof store.getState>
+
 export const persistor = persistStore(store)
 export default store
