+++ src/features/filterContests/ui/filterController.scss
@@ -183,8 +183,6 @@
     }
 }
 
-// SLIDER
-
 .rc-slider {
     .rc-slider-rail {
         background-color: var(--grey);
