+++ src/shared/ui/icon/icon.tsx
@@ -30,7 +30,6 @@ export default function Icon(props: IIcon) {
             width={width}
             height={height}
             className={clsx('icon', className)}
-            // eslint-disable-next-line react/jsx-props-no-spreading
             {...rest}
         />
     )
